// apps/api/src/sys/brevo/types.ts
export interface ErrorModel {
  /**
   * Error code displayed in case of a failure
   * @example "method_not_allowed"
   */
  code:
    | "invalid_parameter"
    | "missing_parameter"
    | "out_of_range"
    | "campaign_processing"
    | "campaign_sent"
    | "document_not_found"
    | "reseller_permission_denied"
    | "not_enough_credits"
    | "permission_denied"
    | "duplicate_parameter"
    | "duplicate_request"
    | "method_not_allowed"
    | "unauthorized"
    | "account_under_validation"
    | "not_acceptable";
  /**
   * Readable message associated to the failure
   * @example "POST Method is not allowed on this path"
   */
  message: string;
}
// ...
export interface CreateUpdateContactModel {
  /**
   * ID of the contact when a new contact is created
   * @format int64
   * @example 122
   */
  id?: number;
}
// ...
export interface CreateSmtpEmail {
  /**
   * Message ID of the transactional email sent
   * @example "<201798300811.5787683@relay.domain.com>"
   */
  messageId: string;
}
// ...
export interface GetContactDetails {
  /**
   * Email address of the contact for which you requested the details
   * @format email
   * @example "john.smith@example.com"
   */
  email: string;
  /**
   * ID of the contact for which you requested the details
   * @format int64
   * @example 32
   */
  id: number;
  /**
   * Blacklist status for email campaigns (true=blacklisted, false=not blacklisted)
   * @example false
   */
  emailBlacklisted: boolean;
  /**
   * Blacklist status for SMS campaigns (true=blacklisted, false=not blacklisted)
   * @example true
   */
  smsBlacklisted: boolean;
  /**
   * Creation UTC date-time of the contact (YYYY-MM-DDTHH:mm:ss.SSSZ)
   * @format date-time
   * @example "2017-05-12T12:30:00Z"
   */
  createdAt: string;
  /**
   * Last modification UTC date-time of the contact (YYYY-MM-DDTHH:mm:ss.SSSZ)
   * @format date-time
   * @example "2017-05-12T12:30:00Z"
   */
  modifiedAt: string;
  listIds: number[];
  listUnsubscribed?: number[];
  /**
   * Set of attributes of the contact
   * @example {"name":"Joe","email":"joe@example.com"}
   */
  attributes: object;
}

export type GetExtendedContactDetails = GetContactDetails & {
  /** Campaign statistics of the contact */
  statistics: {
    /** Listing of the sent campaign for the contact */
    messagesSent?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /**
       * UTC date-time of the event
       * @format date-time
       * @example "2017-03-12T20:15:13Z"
       */
      eventTime: string;
    }[];
    /** Listing of the hardbounes generated by the contact */
    hardBounces?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /**
       * UTC date-time of the event
       * @format date-time
       * @example "2017-03-12T20:15:13Z"
       */
      eventTime: string;
    }[];
    /** Listing of the softbounes generated by the contact */
    softBounces?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /**
       * UTC date-time of the event
       * @format date-time
       * @example "2017-03-12T20:15:13Z"
       */
      eventTime: string;
    }[];
    /** Listing of the complaints generated by the contact */
    complaints?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /**
       * UTC date-time of the event
       * @format date-time
       * @example "2017-03-12T20:15:13Z"
       */
      eventTime: string;
    }[];
    /** Listing of the unsubscription for the contact */
    unsubscriptions?: {
      /** Contact unsubscribe via unsubscription link in a campaign */
      userUnsubscription: {
        /**
         * ID of the campaign which generated the event
         * @format int64
         * @example 3
         */
        campaignId: number;
        /**
         * UTC date-time of the event
         * @format date-time
         * @example "2017-03-12T20:15:13Z"
         */
        eventTime: string;
        /**
         * IP from which the user has unsubscribed
         * @example "165.87.3.15"
         */
        ip?: string;
      }[];
      /** Contact has been unsubscribed from the administrator */
      adminUnsubscription: {
        /**
         * UTC date-time of the event
         * @format date-time
         * @example "2017-03-12T20:15:13Z"
         */
        eventTime: string;
        /**
         * IP from which the user has been unsubscribed
         * @example "165.87.3.15"
         */
        ip?: string;
      }[];
    };
    /** Listing of the openings generated by the contact */
    opened?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /**
       * Number of openings for the campaign
       * @format int64
       * @example 1
       */
      count: number;
      /**
       * UTC date-time of the event
       * @format date-time
       * @example "2017-03-12T20:15:13Z"
       */
      eventTime: string;
      /**
       * IP from which the user has opened the email
       * @example "165.87.3.15"
       */
      ip: string;
    }[];
    /** Listing of the clicks generated by the contact */
    clicked?: {
      /**
       * ID of the campaign which generated the event
       * @format int64
       * @example 3
       */
      campaignId: number;
      /** Listing of the clicked links for the campaign */
      links: {
        /**
         * Number of clicks on this link for the campaign
         * @format int64
         * @example 1
         */
        count: number;
        /**
         * UTC date-time of the event
         * @format date-time
         * @example "2017-03-12T20:15:13Z"
         */
        eventTime: string;
        /**
         * IP from which the user has clicked on the link
         * @example "165.87.3.15"
         */
        ip: string;
        /**
         * URL of the clicked link
         * @example "www.myshop.com"
         */
        url: string;
      }[];
    }[];
    /** Listing of the transactional attributes for the contact */
    transacAttributes?: object[];
  };
};
// ...
export interface HttpResponse<D extends unknown, E extends unknown = unknown>
  extends Response {
  data: D;
  error: E;
}